requirements.txt
allure-behave==2.14.0
Appium-Python-Client==5.0.0
behave==1.2.6
asserts==0.13.1
python-dotenv==1.1.0